+++ CHANGE_LOG.md
@@ -1,3 +1,5 @@
+- 2023/11/26 - `ready.js` - Added reminder in case the direct message fails
+- 2023/11/26 - `settings.js` - Added `reminderChannel` 
 - 2023/11/26 - `ready.js` - Added *Auto-cancel Events* interval and *Remind Players* interval. These need to be tested.
 - 2023/11/26 - `interactionCreate.js` & `helpers.js` - Buttons now correctly change an event's status and send the appropriate message.
 - 2023/11/25 - `create-session.js` - Reworked embeds to show Game Master to handle multiple GMs per server in teh future if needed.